refactor(fetch): renamed fetch_finary to source
