+++ finalynx/fetch/__init__.py
@@ -18,5 +18,5 @@ import finary_uapi.constants  # type: ignore # Set the credentials and cookies c
 finary_uapi.constants.CREDENTIAL_FILE = os.path.join(os.path.dirname(__file__), "finary_credentials.json")
 finary_uapi.constants.COOKIE_FILENAME = os.path.join(os.path.dirname(__file__), "finary_cookies.txt")
 
-from .fetch_finary import FetchFinary
+from .source_finary import FetchFinary
 from ..console import console
